move TestCheckSpender to data/transactions
